+++ include/linux/boost_control.h
@@ -17,7 +17,7 @@ extern bool enable_fb_boost;
 extern u32 input_boost_freq_perf;
 extern u32 input_boost_freq_lp;
 extern u32 wake_boost_duration;
-extern u32 cpubw_boost_freq;
+extern unsigned long cpubw_boost_freq;
 extern u32 cpubw_wake_boost_duration;
 extern u16 dynamic_stune_boost;
 extern u16 app_launch_boost_ms;
